Strip dashes from accountPin for Verizon donors (#92)
### Rationale

We have recently learned that Verizon will display their users account
PINs with dashes (e.g. 123-456) for better readability, but actually
requires the PIN to be sent without dashes for port outs.
Naturally, this cause users to copy the PIN in the format provided by
Verizon, causing failed portings.

We have adapted [Connect](https://gigs.com/connect) to automatically
strip dashes from the accountPin field in the porting form for Verizon
portings and decided to backport this functionality to our other
products as well.

### Changes
`<StepCarrierDetailsForm />` defines a new function (outside of the
component), `normalizeAccountPin`, which handles the removal of dashes
in case the `donorProvider` is "Verizon". The trimming that was done in
`onSubmit` before has been moved into this function as well.